+++ src/main/java/learn/thyme/thyteams/infrastructure/validation/ValidationGroupSequence.java
@@ -1,8 +1,0 @@
-package learn.thyme.thyteams.infrastructure.validation;
-
-import javax.validation.GroupSequence;
-import javax.validation.groups.Default;
-
-@GroupSequence({Default.class, ValidationGroupOne.class, ValidationGroupTwo.class})
-public interface ValidationGroupSequence {
-}