+++ src/main/java/learn/thyme/thyteams/infrastructure/web/EditMode.java
@@ -1,0 +1,6 @@
+package learn.thyme.thyteams.infrastructure.web;
+
+public enum EditMode {
+    CREATE,
+    UPDATE
+}