+++ src/main/java/learn/thyme/thyteams/infrastructure/web/GlobalControllerAdvice.java
@@ -1,0 +1,25 @@
+package learn.thyme.thyteams.infrastructure.web;
+
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.http.HttpStatus;
+import org.springframework.orm.ObjectOptimisticLockingFailureException;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.servlet.ModelAndView;
+
+import javax.servlet.http.HttpServletRequest;
+
+@ControllerAdvice
+public class GlobalControllerAdvice {
+
+    @ResponseStatus(HttpStatus.CONFLICT)
+    @ExceptionHandler({DataIntegrityViolationException.class,
+            ObjectOptimisticLockingFailureException.class})
+    public ModelAndView handleConflicts(HttpServletRequest request, Exception e) {
+        ModelAndView result = new ModelAndView("error/409");
+        // Add the request URL as url in the model so the error page can use this.
+        result.addObject("url", request.getRequestURL());
+        return result;
+    }
+}