+++ src/main/java/learn/thyme/thyteams/user/User.java
@@ -1,6 +1,6 @@
 package learn.thyme.thyteams.user;
 
-import io.github.wimdeblauwe.jpearl.AbstractEntity;
+import io.github.wimdeblauwe.jpearl.AbstractVersionedEntity;
 
 import javax.persistence.Entity;
 import javax.persistence.EnumType;
@@ -11,7 +11,7 @@ import java.time.LocalDate;
 
 @Entity
 @Table(name = "tt_user")
-public class User extends AbstractEntity<UserId> {
+public class User extends AbstractVersionedEntity<UserId> {
 
     @NotNull
     private UserName userName;
@@ -59,5 +59,25 @@ public class User extends AbstractEntity<UserId> {
 
     public PhoneNumber getPhoneNumber() {
         return phoneNumber;
+    }
+
+    public void setUserName(UserName userName) {
+        this.userName = userName;
+    }
+
+    public void setGender(Gender gender) {
+        this.gender = gender;
+    }
+
+    public void setBirthday(LocalDate birthday) {
+        this.birthday = birthday;
+    }
+
+    public void setEmail(Email email) {
+        this.email = email;
+    }
+
+    public void setPhoneNumber(PhoneNumber phoneNumber) {
+        this.phoneNumber = phoneNumber;
     }
 }