+++ src/main/java/learn/thyme/thyteams/user/UserService.java
@@ -3,8 +3,14 @@ package learn.thyme.thyteams.user;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
+import java.util.Optional;
+
 public interface UserService {
     User createUser(CreateUserParameters parameters);
+
+    Optional<User> getUser(UserId userId);
+
+    User editUser(UserId userId, EditUserParameters parameters);
 
     Page<User> getUsers(Pageable pageable);