+++ src/main/java/learn/thyme/thyteams/user/UserServiceImpl.java
@@ -2,8 +2,11 @@ package learn.thyme.thyteams.user;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.orm.ObjectOptimisticLockingFailureException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.Optional;
 
 @Service
 @Transactional
@@ -27,11 +30,30 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
+    @Transactional(readOnly = true)
+    public Optional<User> getUser(UserId userId) {
+        return repository.findById(userId);
+    }
+
+    @Override
+    public User editUser(UserId userId, EditUserParameters parameters) {
+        User user = repository.findById(userId)
+                .orElseThrow(() -> new UserNotFoundException(userId));
+        if (parameters.getVersion() != user.getVersion()) {
+            throw new ObjectOptimisticLockingFailureException(User.class, user.getId().asString());
+        }
+        parameters.update(user); // Saving the updated user is done automatically by JPA/Hibernate.
+        return user;
+    }
+
+    @Override
+    @Transactional(readOnly = true)
     public Page<User> getUsers(Pageable pageable) {
         return repository.findAll(pageable);
     }
 
     @Override
+    @Transactional(readOnly = true)
     public boolean userWithEmailExists(Email email) {
         return repository.existsByEmail(email);
     }