+++ src/main/java/learn/thyme/thyteams/user/web/CreateUserValidationGroupSequence.java
@@ -1,0 +1,11 @@
+package learn.thyme.thyteams.user.web;
+
+import learn.thyme.thyteams.infrastructure.validation.ValidationGroupOne;
+import learn.thyme.thyteams.infrastructure.validation.ValidationGroupTwo;
+
+import javax.validation.GroupSequence;
+import javax.validation.groups.Default;
+
+@GroupSequence({Default.class, ValidationGroupOne.class, ValidationGroupTwo.class})
+public interface CreateUserValidationGroupSequence {
+}