+++ src/main/java/learn/thyme/thyteams/user/web/EditUserValidationGroupSequence.java
@@ -1,0 +1,10 @@
+package learn.thyme.thyteams.user.web;
+
+import learn.thyme.thyteams.infrastructure.validation.ValidationGroupOne;
+
+import javax.validation.GroupSequence;
+import javax.validation.groups.Default;
+
+@GroupSequence({Default.class, ValidationGroupOne.class})
+public interface EditUserValidationGroupSequence {
+}